+++ src/Armada.Server/MissionLandingHandler.cs
@@ -483,6 +483,8 @@ namespace Armada.Server
                 {
                     mission.Status = MissionStatusEnum.Complete;
                     mission.CompletedUtc = DateTime.UtcNow;
+                    if (mission.StartedUtc.HasValue)
+                        mission.TotalRuntimeSeconds = (mission.CompletedUtc.Value - mission.StartedUtc.Value).TotalSeconds;
                     mission.LastUpdateUtc = DateTime.UtcNow;
                     await _Database.Missions.UpdateAsync(mission).ConfigureAwait(false);
                     _Logging.Info(_Header + "mission " + mission.Id + " landed successfully, status set to Complete");
@@ -624,6 +626,8 @@ namespace Armada.Server
                 {
                     mission.Status = MissionStatusEnum.Complete;
                     mission.CompletedUtc = DateTime.UtcNow;
+                    if (mission.StartedUtc.HasValue)
+                        mission.TotalRuntimeSeconds = (mission.CompletedUtc.Value - mission.StartedUtc.Value).TotalSeconds;
                     mission.LastUpdateUtc = DateTime.UtcNow;
                     await _Database.Missions.UpdateAsync(mission).ConfigureAwait(false);
                     _Logging.Info(_Header + "PR reconciler: mission " + mission.Id + " PR merged, status set to Complete");
@@ -681,6 +685,8 @@ namespace Armada.Server
                             {
                                 mission.Status = MissionStatusEnum.Complete;
                                 mission.CompletedUtc = DateTime.UtcNow;
+                                if (mission.StartedUtc.HasValue)
+                                    mission.TotalRuntimeSeconds = (mission.CompletedUtc.Value - mission.StartedUtc.Value).TotalSeconds;
                                 mission.LastUpdateUtc = DateTime.UtcNow;
                                 await _Database.Missions.UpdateAsync(mission).ConfigureAwait(false);
                                 _Logging.Info(_Header + "mission " + missionId + " PR merged, status set to Complete");